zerofill: Change the type of len argument of glfs_zerofill() to off_t
glfs_zerofill() can be potentially called to zero-out entire file and
hence allow for bigger value of length parameter.

BUG: 1028673
Reviewed-on: http://review.gluster.org/6266
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Tested-by: M. Mohan Kumar <mohan@in.ibm.com>

+++ libglusterfs/src/defaults.h
@@ -259,7 +259,7 @@ int32_t default_zerofill(call_frame_t *frame,
                         xlator_t *this,
                         fd_t *fd,
                         off_t offset,
-                        size_t len, dict_t *xdata);
+                        off_t len, dict_t *xdata);
 
 
 /* Resume */
@@ -488,7 +488,7 @@ int32_t default_zerofill_resume(call_frame_t *frame,
                                xlator_t *this,
                                fd_t *fd,
                                off_t offset,
-                               size_t len, dict_t *xdata);
+                               off_t len, dict_t *xdata);
 
 
 /* _cbk */
