+++ xlators/features/marker/src/marker.c
@@ -1991,7 +1991,7 @@ out:
 
 int32_t
 marker_zerofill(call_frame_t *frame, xlator_t *this, fd_t *fd, off_t offset,
-               size_t len, dict_t *xdata)
+               off_t len, dict_t *xdata)
 {
         int32_t          ret   = 0;
         marker_local_t  *local = NULL;
